second commit of weibo part
