+++ core/utils.py
@@ -172,53 +172,3 @@ class Status:
 
 
 twitter = TwitterAPI(CONSUMER_KEY, CONSUMER_SECRET, ACCESS_TOKEN_KEY, ACCESS_TOKEN_SECRET, tweet_mode='extended')
-
-
-def base62_encode(num, alphabet='0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ'):
-    """Encode a number in Base X
-
-    `num`: The number to encode
-    `alphabet`: The alphabet to use for encoding
-    """
-    if num == 0:
-        return alphabet[0]
-    arr = []
-    base = len(alphabet)
-    while num:
-        rem = num % base
-        num = num // base
-        arr.append(alphabet[rem])
-    arr.reverse()
-    return ''.join(arr)
-
-
-def mid_to_url(midint):
-    midint = str(midint)[::-1]
-    size = int(len(midint) // 7) if len(midint) % 7 == 0 else int(len(midint) // 7) + 1
-    result = []
-    for i in range(size):
-        s = midint[i * 7: (i + 1) * 7][::-1]
-        s = base62_encode(int(s))
-        s_len = len(s)
-        if i < size - 1 and len(s) < 4:
-            s = '0' * (4 - s_len) + s
-        result.append(s)
-    result.reverse()
-    return ''.join(result)
-
-
-class WeiboPost:
-    def __init__(self, status):
-        self.status = status
-
-    @property
-    def id(self):
-        return self.status['id']
-
-    @property
-    def text(self):
-        return self.status['text']
-
-    @property
-    def user(self):
-        return self.status['user']
